chore(docs):update readme * Changed the project name to protractor-cli * Updated readme with protracto-cli info * Moved the spec files to test folder
